// DXC_Frontend/src/api/models/bannerDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface BannerDto {
  id?: number
  publicId?: string
  /** @nullable */
  title: string | null
  imagePublicId?: string
  /** @nullable */
  position: string | null
  /** @nullable */
  bannerType: string | null
  /** @nullable */
  nativeParams?: string | null
  /** @nullable */
  webLink?: string | null
  thuTu?: number
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
}